move Context helpers to utils
