+++ scenario/context/CometContext.ts
@@ -13,7 +13,7 @@ import {
 } from '../constraints';
 import CometActor from './CometActor';
 import CometAsset from './CometAsset';
-import { ProposalState, OpenProposal } from './Gov';
+import { OpenProposal } from './Gov';
 import {
   CometInterface,
   ERC20__factory,
@@ -30,8 +30,7 @@ import { sourceTokens } from '../../plugins/scenario/utils/TokenSourcer';
 import { ProtocolConfiguration, deployComet, COMP_WHALES } from '../../src/deploy';
 import { AddressLike, getAddressFromNumber, resolveAddress } from './Address';
 import { Requirements } from '../constraints/Requirements';
-import { DeploymentManager } from '../../plugins/deployment_manager';
-import { impersonateAddress } from '../../plugins/scenario/World';
+import { executeOpenProposal, fastGovernanceExecute, mineBlocks, setNextBaseFeeToZero, setNextBlockTimestamp } from '../utils';
 
 export type ActorMap = { [name: string]: CometActor };
 export type AssetMap = { [name: string]: CometAsset };
@@ -46,98 +45,6 @@ export interface CometProperties {
   governor: IGovernorBravo;
   rewards: CometRewards;
   bulker: Bulker;
-}
-
-export async function setNextBaseFeeToZero(dm: DeploymentManager) {
-  await dm.hre.network.provider.send('hardhat_setNextBlockBaseFeePerGas', ['0x0']);
-}
-
-export async function mineBlocks(dm: DeploymentManager, blocks: number) {
-  await dm.hre.network.provider.send('hardhat_mine', [`0x${blocks.toString(16)}`]);
-}
-
-export async function setNextBlockTimestamp(dm: DeploymentManager, timestamp: number) {
-  await dm.hre.ethers.provider.send('evm_setNextBlockTimestamp', [timestamp]);
-}
-
-export async function executeOpenProposal(
-  dm: DeploymentManager,
-  { id, startBlock, endBlock }: OpenProposal
-) {
-  const governor = await dm.contract('governor');
-
-  if (!governor) {
-    throw new Error("cannot execute proposal without governor");
-  }
-
-  const blockNow = await dm.hre.ethers.provider.getBlockNumber();
-  const blocksUntilStart = startBlock - blockNow;
-  const blocksUntilEnd = endBlock - Math.max(startBlock, blockNow);
-
-  if (blocksUntilStart > 0) {
-    await mineBlocks(dm, blocksUntilStart);
-  }
-
-  if (blocksUntilEnd > 0) {
-    for (const whale of COMP_WHALES) {
-      try {
-        // Voting can fail if voter has already voted
-        const voter = await impersonateAddress(dm, whale);
-        await setNextBaseFeeToZero(dm);
-        await governor.connect(voter).castVote(id, 1, { gasPrice: 0 });
-      } catch (err) {
-        debug(`Error while voting for ${whale}`, err.message);
-      }
-    }
-    await mineBlocks(dm, blocksUntilEnd);
-  }
-
-  // Queue proposal (maybe)
-  const state = await governor.state(id);
-  if (state == ProposalState.Succeeded) {
-    await setNextBaseFeeToZero(dm);
-    await governor.queue(id, { gasPrice: 0 });
-  }
-
-  const proposal = await governor.proposals(id);
-  await setNextBlockTimestamp(dm, proposal.eta.toNumber() + 1);
-
-  // Execute proposal (w/ gas limit so we see if exec reverts, not a gas estimation error)
-  await setNextBaseFeeToZero(dm);
-  await governor.execute(id, { gasPrice: 0, gasLimit: 12000000 });
-}
-
-// Instantly executes some actions through the governance proposal process
-export async function fastGovernanceExecute(
-  dm: DeploymentManager,
-  proposer: SignerWithAddress,
-  targets: string[],
-  values: BigNumberish[],
-  signatures: string[],
-  calldatas: string[]
-) {
-  const governor = await dm.contract('governor');
-
-  if (!governor) {
-    throw new Error("cannot create governance proposal without governor");
-  }
-
-  await setNextBaseFeeToZero(dm);
-
-  const proposeTxn = await (
-    await governor.connect(proposer).propose(
-      targets,
-      values,
-      signatures,
-      calldatas,
-      'FastExecuteProposal',
-      { gasPrice: 0 }
-    )
-  ).wait();
-  const proposeEvent = proposeTxn.events.find(event => event.event === 'ProposalCreated');
-  const [id, , , , , , startBlock, endBlock] = proposeEvent.args;
-
-  await executeOpenProposal(dm, { id, startBlock, endBlock });
 }
 
 export class CometContext {
